+++ resources/views/recruiter/show.blade.php
@@ -3,11 +3,27 @@
 @section('content')
 <div class="container">
     <div class="row">
-        <div class="col-md-8 col-md-offset-2">
+        <div class="col-md-10 col-md-offset-1 custom-padding">
             <div class="panel panel-default">
-                <div class="panel-heading"><a href="/recruiter">Recruiters</a> <span class="fa fa-chevron-right"></span> {{ $recruiter->first_name }} {{ $recruiter->last_name }}
-                    <a href="{{ route('recruiter.edit', $recruiter->id) }}" class="btn btn-primary btn-sm pull-right">Edit </a>
-                </div>
+                <?php
+                    if (Auth::guest() != true){
+                        if (Auth::user()->user_type_id == 4){
+                            ?>
+                                <div class="panel-heading"><a href="/recruiter">Recruiters</a> <span class="fa fa-chevron-right"></span> {{ $recruiter->first_name }} {{ $recruiter->last_name }}
+                                    <a href="{{ route('recruiter.edit', $recruiter->id) }}" class="btn btn-primary btn-sm pull-right">Edit </a>
+                                </div>
+                            <?php
+                        }
+                        else{
+                            ?>
+                                <div class="panel-heading">My Profile
+                                    <a href="{{ route('recruiter.edit', $recruiter->id) }}" class="btn btn-primary btn-sm pull-right">Edit </a>
+                                </div>
+                            <?php
+                        }
+                    }
+                 ?>
+
                 <div class="panel-body">
                     <div>
                         <table class="table table-striped table-responsive">
@@ -77,21 +93,25 @@
                         </table>
                     </div>
                 </div>
-                <div class="panel-footer">
+                <?php
+                    if (Auth::guest() != true){
+                        if (Auth::user()->user_type_id == 4){
+                            ?>
+                                <div class="panel-footer">
+                                    {!! Form::open([
+                                        'route' => ['recruiter.destroy', $recruiter->id],
+                                        'method' => 'delete',
+                                        'class' => 'form-horizontal'
+                                    ]) !!}
 
-                    {!! Form::open([
-                        'route' => ['recruiter.destroy', $recruiter->id],
-                        'method' => 'delete',
-                        'class' => 'form-horizontal'
-                    ]) !!}
+                                    <button type="submit" class="btn btn-danger">Delete</button>
 
-                    <button type="submit" class="btn btn-danger">
-                        Delete
-                    </button>
-
-                    {!! Form::close() !!}
-
-                </div>
+                                    {!! Form::close() !!}                
+                                </div>
+                            <?php
+                        }
+                    }
+                 ?>
             </div>
         </div>
     </div>
